Add files via upload
This is my basketball scoreboard project that I used for learning HTML, CSS and Javascript. It has basic functions which can add 1, 2 or 3 points for Home or Guest team. I also implemented an undo function which reverts the score to the previous one. I did that by storing each state in an array. Also when score is added to the scoreboard, an array index is updated. So when "Undo" function is called it always gets the last score that was on the scoreboard.